Splitting out some files
Again, this does NOT need to be a monorepo, this is more of an excuse to
learn how workspaces work.
