+++ docs/index.html
@@ -2,10 +2,10 @@
 <html lang="en">
   <head>
     <meta charset="UTF-8" />
-    <link rel="icon" type="image/svg+xml" href="/vite.svg" />
+    <link rel="icon" type="image/svg+xml" href="/quickshop/favicon.ico" />
     <meta name="viewport" content="width=device-width, initial-scale=1.0" />
-    <title>Vite + React + TS</title>
-    <script type="module" crossorigin src="/quickshop/assets/index-6wCDmp3d.js"></script>
+    <title>QuickShop</title>
+    <script type="module" crossorigin src="/quickshop/assets/index-IsLuLLAL.js"></script>
     <link rel="stylesheet" crossorigin href="/quickshop/assets/index-fEkTNaLb.css">
   </head>
   <body>
