+++ packages/quickshop-core/src/types.ts
@@ -1,0 +1,13 @@
+export type Item = {
+  label: string;
+  department: string;
+};
+
+export type Department = {
+  label: string;
+};
+
+export type ItemsByDepartment = {
+  departmentName: string;
+  items: Item[];
+};
