+++ tsconfig.json
@@ -4,7 +4,6 @@
       "module": "commonjs",
       "declaration": true,
       "sourceMap": true,
-      "composite": true,
       "strict": true,
       "moduleResolution": "node",
       "esModuleInterop": true,
